Moved LogicalRange to time-data and removed LogicalPoint type
